change float to double

+++ src/Cofactor.java
@@ -1,11 +1,11 @@
 package cofactor;
 public class Cofactor {
     
-    public static float[][] transpose(float[][] m) {
+    public static double[][] transpose(double[][] m) {
         // transpose matrix
         int row = m.length;
         int col = m[0].length;
-        float[][] trans = new float[row][col];
+        double[][] trans = new double[row][col];
 
         for (int i = 0; i < row; i++) {
             for (int j = 0; j < col; j++) {
@@ -17,7 +17,7 @@ public class Cofactor {
     }
 
 
-    public static void getCofactor(float[][] m, float[][] temp, int p, int q) {
+    public static void getCofactor(double[][] m, double[][] temp, int p, int q) {
         // Mendapatkan kofaktor dari m dan menyimpannya di temp
         int n = m.length;
         
@@ -51,13 +51,13 @@ public class Cofactor {
     }
 
 
-    public static float determinant(float[][] m, int n) {
+    public static double determinant(double[][] m, int n) {
         // Mendapatkan determinan dari m secara rekursif menggunakan metode kofaktor
-        float det = 0;
+        double det = 0;
         if (n == 1) {
             return m[0][0];
         }
-        float[][] temp = new float[n][n];
+        double[][] temp = new double[n][n];
         int sign = 1;
         for (int i = 0; i < n; i++) {
             getCofactor(m, temp, 0, i);
@@ -67,16 +67,16 @@ public class Cofactor {
         return det;
     }
 
-    public static float[][] adjoint(float[][] m){
+    public static double[][] adjoint(double[][] m){
         // Mendapatkan matriks adjoin dari m secara rekursif
         int n = m.length;
-        float[][] adj = new float[n][n];
+        double[][] adj = new double[n][n];
         if (n == 1){
             adj[0][0] = 1;
             return adj;
         }
         int sign = 1;
-        float[][] temp = new float[n][n];
+        double[][] temp = new double[n][n];
         for (int i = 0; i < n; i++){
             for (int j = 0; j < n; j++){
                 getCofactor(m, temp, i, j);
@@ -87,16 +87,16 @@ public class Cofactor {
         return adj;
     }
 
-    public static float[][] inverse(float[][] m){
+    public static double[][] inverse(double[][] m){
         // Mendapatkan matriks invers dari m dari adjoint dan determinan
         int n = m.length;
-        float[][] inv = new float[n][n];
-        float det = determinant(m, n);
+        double[][] inv = new double[n][n];
+        double det = determinant(m, n);
         if (det == 0){
             System.out.println("Determinan 0, tidak memiliki invers");
             return inv;
         }
-        float[][] adj = adjoint(m);
+        double[][] adj = adjoint(m);
         for (int i = 0; i < n; i++){
             for (int j = 0; j < n; j++){
                 inv[i][j] = adj[i][j] / det;
